[models,profile]Touch up product_list price to decimal field. Touch up profile page.

+++ main_app/models.py
@@ -29,7 +29,7 @@ class Product(models.Model):
     name = models.CharField(max_length=100)
     description = models.CharField(max_length=500)
     image = models.ImageField()
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
     slug = models.SlugField()
 
     def __str__(self):
@@ -59,9 +59,8 @@ class Order(models.Model):
 class OrderItem(models.Model):
     order = models.ForeignKey(Order, related_name='items', on_delete=models.CASCADE)
     product = models.ForeignKey(Product, related_name='order_items', on_delete=models.CASCADE)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
     quantity = models.IntegerField()
-    # amount = models.IntegerField()
 
     def __str__(self):
         return '{}'.format(self.id)
